Reorganized files
Moved the Python scripts to a new folder, BeautifulSites4, that is to be made package.
